qc/fastq_screen: relocate to 'qc/apps/fastq_screen'.
